db/requirements.txt
cassandra-driver==3.14.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.3
numpy==1.15.0
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
